Reorganize components to panel and actually add global components
