+++ src/components/global/Loadings.tsx
@@ -1,0 +1,7 @@
+export const DefaultPageLoader = () => {
+	return (
+		<main className='flex flex-col justify-items-center h-svh w-svw'>
+			<h1 className='block 4xl text-center'>Loading...</h1>
+		</main>
+	);
+};
