+++ src/components/panel/DualView.tsx
