+++ src/components/panel/DualViewServer.tsx
